Rename Netlify function to .mjs
